% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{betaTilde}
\alias{betaTilde}
\title{Bias Correct OLS}
\usage{
betaTilde(coef, S, X, N)
}
\arguments{
\item{coef}{Vector of OLS coefficients}

\item{S}{Matrix of error terms. Dimensions must be length(coef) x length(coef)}

\item{X}{Design matrix that produced OLS coefficients}

\item{N}{Number of rows in X}
}
\value{
Returns a list:
\item{bias_correct}{Vector of bias corrected coefficients}
\item{x_prime_x}{X'X matrix to be used in variance estimationo}
}
\description{
This function takes OLS coefficients as inputs and bias corrects according to the noise matrix, S.
}
