% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{betaTildeInt}
\alias{betaTildeInt}
\title{Bias Correct OLS model with interaction term}
\usage{
betaTildeInt(coef, S, X, N, int1, int2)
}
\arguments{
\item{coef}{Vector of OLS coefficients}

\item{S}{Matrix of error terms. Dimensions must be length(coef) x length(coef)}

\item{X}{Design matrix that produced OLS coefficients}

\item{N}{Number of rows in X}

\item{int1}{Column index of first interaction variable in X}

\item{int2}{Column index of second interaction variable in X}
}
\value{
Returns a list:
\item{bias_correct}{Vector of bias corrected coefficients}
\item{x_prime_x}{X'X matrix to be used in variance estimationo}
}
\description{
This function takes OLS coefficients from a moel including an
interaction term as inputs and bias corrects according to the noise matrix, S.
}
