% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{covXX}
\alias{covXX}
\title{Estimates \eqn{Cov(X_k'X_j, X_m'X_l)}}
\usage{
covXX(k, j, l, m, S, X, x_prime_x, N)
}
\arguments{
\item{k}{column index of first X.}

\item{j}{column index of second X.}

\item{l}{column index of fourth X.}

\item{m}{column index of third X.}

\item{S}{Matrix of error terms. Dimensions must be ncol(X) x ncol(X)}

\item{X}{Design matrix that produced OLS coefficients}

\item{x_prime_x}{X'X}

\item{N}{Number of rows in X}
}
\value{
Estimate of Cov(X_k'X_j, X_m'X_l)
}
\description{
This function estimates components of variance-covariance matrix to subsequently
draw random variable of \eqn{\tilde{\beta}} from a multivariate normal.
}
