% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{covXyXX}
\alias{covXyXX}
\title{Estimates \eqn{Cov(X_k'y, X_j'X_m)}}
\usage{
covXyXX(k, j, m, S, X, Y, N)
}
\arguments{
\item{k}{column index of first X.}

\item{j}{column index of second X.}

\item{m}{column index of third X.}

\item{S}{Matrix of error terms. Dimensions must be ncol(X) x ncol(X)}

\item{X}{Design matrix that produced OLS coefficients}

\item{Y}{Dependent variable from OLS model}

\item{N}{Number of rows in X}
}
\value{
Estimate of \eqn{Cov(X_k'y, X_j'X_m)}
}
\description{
This function estimates components of variance-covariance matrix to subsequently
draw random variable of \eqn{\tilde{\beta}} from a multivariate normal.
}
