% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{covXyXy}
\alias{covXyXy}
\title{Estimates \eqn{Cov(X_k'y, X_j'y)}}
\usage{
covXyXy(k, j, S, X, Y, sigma_sq, x_prime_x, N)
}
\arguments{
\item{k}{column index of first X.}

\item{j}{column index of second X.}

\item{S}{Matrix of error terms. Dimensions must be ncol(X) x ncol(X)}

\item{X}{Design matrix that produced OLS coefficients}

\item{Y}{Dependent variable from OLS model}

\item{sigma_sq}{estimate of \Sigma^2}

\item{x_prime_x}{X'X}

\item{N}{Number of rows in X}
}
\value{
Estimate of Cov(X_k'y, X_j'y)
}
\description{
This function estimates components of variance-covariance matrix to subsequently
draw random variable of \eqn{\tilde{\beta}} from a multivariate normal.
}
