% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{descriptiveDP}
\alias{descriptiveDP}
\title{Estimate the first R raw moments}
\usage{
descriptiveDP(variable, data)
}
\arguments{
\item{variable}{Variable in data to estimate descriptive statistics of}

\item{data}{Data}
}
\value{
Table containing descriptive stats of private data column (estimated) and DP data column (observed)
}
\description{
This function estimates descriptive statistics of private data from column in DP data
}
