% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diagnostics.R
\name{diagnosticsDP}
\alias{diagnosticsDP}
\title{DP regression diagnostics}
\usage{
diagnosticsDP(lmdp_obj)
}
\arguments{
\item{lmdp_obj}{output from \code{\link{lmdp}}}
}
\value{
Summary tesst of normality and homoskedasticity of errors
}
\description{
Run diagnostic check on DP linear regression
}
