% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distributionDP}
\alias{distributionDP}
\title{Paramaterize distrributions}
\usage{
distributionDP(variable, data, distributions, moments_fit = 6,
  plot = TRUE, plot_dp = TRUE)
}
\arguments{
\item{variable}{Variable name in dataa}

\item{data}{DP data (first row should contain sd of noise added to `variable``)}

\item{distributions}{Vector of distributions to fit. Options include 'Normal', 'Poisson', 'NB', 'ZIP', 'ZINB'}

\item{moments_fit}{Number of moments to fit implied distribution to}

\item{plot_dp}{Include DP in plot}

\item{Plot}{Plot histogram of distributions TRUE/FALSE}
}
\value{
Returns list of paramater estimates:
}
\description{
Paramaterize distributions for private data based on moments estimated from DP data
}
