% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{estZIP}
\alias{estZIP}
\title{Zero Inflated Poisson paramaters}
\usage{
estZIP(moments)
}
\arguments{
\item{moments}{vector of moments}
}
\value{
Vector of ZIP paramaters
}
\description{
Estimates paramaters of a ZIP using moments
}
