% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{estZNB}
\alias{estZNB}
\title{Zero-Inflated NegBin paramaters}
\usage{
estZNB(moments)
}
\arguments{
\item{moments}{vector of moments}
}
\value{
Vector of ZNB paramaters
}
\description{
Estimates paramaters of a ZINB using estimated moments
}
