% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{genData}
\alias{genData}
\title{Simulate data}
\usage{
genData(N)
}
\arguments{
\item{N}{Number of observations in simulated data}

\item{coef}{Vector of OLS coefficients}
}
\value{
Returns a list:
\item{original_data}{Data before DP noise added}
\item{data_errror}{Data after DP noise added }
}
\description{
This function simulatess example data.
}
