% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{genData2}
\alias{genData2}
\title{Simulate data 2}
\usage{
genData2(N)
}
\arguments{
\item{N}{Number of observations in simulated data}
}
\value{
Returns a list:
\item{original_data}{Data before DP noise added}
\item{data_errror}{Data after DP noise added }
}
\description{
This function simulates example data with heteroskedastic error and zero inflated covariates.
}
