% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{lagHerm}
\alias{lagHerm}
\title{Estimate lagged value of column}
\usage{
lagHerm(x, k)
}
\arguments{
\item{x}{vector to lag}

\item{k}{length of lag}
}
\value{
vector of lags
}
\description{
This function estimates first R moments of a column of the private data according to the DP data and the noise matrix, S.
}
