% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{lmdp}
\alias{lmdp}
\title{OLS differential privacy}
\usage{
lmdp(formula, data, bootstrap_var = FALSE, nsims_var = 500,
  noise = NULL)
}
\arguments{
\item{formula}{An \code{\link{lm}} style formula.}

\item{data}{The data to estimate the model on. The first row of data should contain the DP
standard error associated with that column unless noise argument is not NULL.}

\item{bootstrap_var}{If FALSE, then the variance is estimated via simulation.
If TRUE then the variance covariance matrix is estimated via bootstrap methods. Default is
FALSE unless model contains interaction terms/squared terms or fewer than 10000 obsesrvations}

\item{nsims_var}{Number of bootstrap samples/simulations. Default is 500}

\item{noise}{Set a default differentially private standard error for every column of the data matrix}
}
\value{
Returns an object of class lmdp containing:
\item{b}{Inconsistent OLS coefficient estimate}
\item{b_vcov}{Estimate of variance covariance of b}
\item{beta_tilde}{Consistent estimates of coefficients, \eqn{\tilde{\beta}}}
\item{beta_tilde_vcov}{Estimate of variance covariance of \eqn{\tilde{\beta}}}
\item{var_sims}{Full set of simulated/bootstrap estimates of b and \eqn{\tilde{\beta}}}
\item{Sigma_sq_hat}{Estimate of \eqn{\sigma^2}}
\item{vc_pos_def}{Indicator variable = 1 if covariance estimate was PD. NA if bootstrap used}
\item{boot}{Indicator variable = 1 if bootstrap was used to estimate variance}
\item{est_vc}{Variance-covariance matrix used in variance simulation. NA if bootstrap used}
}
\description{
This function works similarly to \code{\link{lm}}. It takes a formula and data and
returns an lmdp object containing bias corrected OLS coefficients and standard errors. The output can be summarised
by inputting it as an argument to \code{\link{summary.lmdp}}. See an overview at \url{bit.ly/PrivUexample}.
}
\examples{
\dontrun{data(dp_data)}
\dontrun{lmdp_test <- lmdp(Y ~ X1 + X2 + X3, data = dp_data)}
\dontrun{summary(lmdp_test)}
}
