% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{momentR}
\alias{momentR}
\title{Estimate the Rth raw moment}
\usage{
momentR(X, S, r)
}
\arguments{
\item{X}{Column from DP data}

\item{S}{Scalar value of DP error.}

\item{r}{Moment to be estimate}
}
\value{
Estimate of Rth moment
}
\description{
This function estimates the Rth moment of a column of the private data according to the DP data and the noise matrix, S.
}
