% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{paramsNormal}
\alias{paramsNormal}
\title{Normal fit}
\usage{
paramsNormal(X, S, R, moments_df, plot_dp)
}
\arguments{
\item{X}{dp vector}

\item{S}{dp noise added to obtain X}

\item{R}{numbers of implied moments to fit}

\item{moments_df}{estimated moments of Z}

\item{plot_dp}{TRUE/FALSE value denoting whether to generate plot}
}
\value{
Returns list of estimates:
\item{params}{Method of moment estimates of mean and variance of normal distribution}
\item{moment_df}{Moments used to estimate params}
\item{moment_fit}{Ratio of estimated moments to implied moments from paramaterized distribution}
\item{moment_precision}{Ratio of moment mean to moment standard error}
\item{plot}{Plot of fitted distribution}
}
\description{
Estimate via method of moments fit and paramaters of a normal distribution from DP data
}
