% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plotDist}
\alias{plotDist}
\title{Distribution plots}
\usage{
plotDist(X, Z_est, plot_dp)
}
\arguments{
\item{X}{dp vector}

\item{Z_est}{Estimate of private vector}
}
\value{
Histogram of dp vector and estimated private vector
}
\description{
Distribution plots
}
