% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{propDP}
\alias{propDP}
\title{Proportions variable}
\usage{
propDP(data, numerator_var, denominator_vars)
}
\arguments{
\item{data}{DP data frame where first row is sd of dp noise}

\item{numerator_var}{A character vector with the names of count variables that make up the weights}

\item{denominator_vars}{A character vector with the names of count variables that make up the weights}
}
\value{
A new weighted average variable that can be added as a new column to the DP data frame
}
\description{
This function creates a proportion vector that can be added as a column to dp data with appropriate standard deviation in first row for use in lmdp
}
