% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{Rmoments}
\alias{Rmoments}
\title{Estimate the first R raw moments}
\usage{
Rmoments(X, S, R)
}
\arguments{
\item{X}{Column from DP data}

\item{S}{Scalar value of DP error.}

\item{R}{First R moments to be estimated}
}
\value{
Vector of estimates of first R moments
}
\description{
This function estimates first R moments of a column of the private data according to the DP data and the noise matrix, S.
}
