% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{scaledMoments}
\alias{scaledMoments}
\title{Estimate the Scaled Moments}
\usage{
scaledMoments(moments)
}
\arguments{
\item{moments}{Vector of estimated raw moments}
}
\value{
Vector of scaled moments
}
\description{
This function estimates mean, variance, skew and kurtosis of the private data
}
