% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sigmaSq}
\alias{sigmaSq}
\title{Estimates \eqn{\sigma^2} from regression}
\usage{
sigmaSq(Y, bias_correct, S, X)
}
\arguments{
\item{Y}{Dependent variable from OLS model}

\item{S}{Matrix of error terms. Dimensions must be length(coef) x length(coef)}

\item{X}{Design matrix that produced OLS coefficients}

\item{bias_corrrect}{Bias corrected OLS coefficients}
}
\value{
Estimate of \eqn{\sigma^2}
}
\description{
This function takes OLS coefficients and a Y vector to estimate \eqn{\sigma^2}
for the model.
}
