% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{simulateEst}
\alias{simulateEst}
\title{Draw re-sampled estimates}
\usage{
simulateEst(X, mu, Sigma, S, N)
}
\arguments{
\item{X}{Design matrix that produced OLS coefficients}

\item{mu}{Mean vector for multivariate normal}

\item{Sigma}{Variance covariance matrix for mjultivariate normal}

\item{S}{Matrix of error terms. Dimensions must be ncol(X) x ncol(X)}

\item{N}{Number of rows in X}
}
\value{
One draw of vector of re-sample b and \eqn{\tilde{\beta}}
}
\description{
This function draws random variables vec(X'X, X'y) from a multivariate normal
and constructs a simulated \eqn{\tilde{\beta}} and b.
}
