% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{summary.lmdp}
\alias{summary.lmdp}
\title{Summary lmdp}
\usage{
\method{summary}{lmdp}(lmdp_object)
}
\arguments{
\item{lmdp_object}{Output from \code{\link{lmdp}}}
}
\value{
Table of consistent coefficient estimates and corrresponding standard errors, t-stats
and p-values.
}
\description{
This function takes an lmdp object as input and produces a summary table comparable to
a summary of an \code{\link[base]{lm}} object.
}
