% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{varianceBoot}
\alias{varianceBoot}
\title{Bootstrap variance}
\usage{
varianceBoot(Y, X, S, nsims, N, int_vars, sq_vars, int1, int2, index_sq)
}
\arguments{
\item{Y}{Dependent variable from OLS model}

\item{X}{Design matrix that produced OLS coefficients}

\item{S}{Matrix of error terms. Dimensions must be ncol(X) x ncol(X)}

\item{nsims}{Number of times to draw from a MVN}

\item{int_vars}{Vector of interaction variables}

\item{sq_vars}{Vector of squared variables}

\item{int1}{Column index of first interaction variable}

\item{int2}{Column index of second interaction variable}

\item{index_sq}{Column index of second interaction variable}
}
\value{
Bootstrap samples of b and \eqn{\tilde{\beta}}
}
\description{
This function takes repeated samples from the data to estimate a bootstrap
distribution of \eqn{\tilde{\beta}} and b.
}
