% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{varianceMVN}
\alias{varianceMVN}
\title{Simulate variance}
\usage{
varianceMVN(X, Y, sigma_sq, nsims, coef, S, N, x_prime_x, x_y)
}
\arguments{
\item{X}{Design matrix that produced OLS coefficients}

\item{Y}{Dependent variable from OLS model}

\item{sigma_sq}{Estimate of \eqn{\sigma^2}}

\item{nsims}{Number of times to draw from a MVN}

\item{coef}{Vector of OLS coefficientss}

\item{S}{Matrix of error terms. Dimensions must be length(coef) x length(coef)}

\item{N}{Number of rows in X}

\item{x_prime_x}{X'X}

\item{x_y}{X'y}
}
\value{
Returns a list:
\item{sims}{Draws of b and \eqn{\tilde{\beta}}}
\item{Sigma}{Estimate of variance-covariance that was used for simulation}
}
\description{
This function takes repeated draws from a MVN to simulate the approximate
distribution of \eqn{\tilde{\beta}} and b.
}
