% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{waDP}
\alias{waDP}
\title{Weighted average variable}
\usage{
waDP(data, weight_vars, values)
}
\arguments{
\item{data}{DP data frame where first row is sd of dp noise}

\item{weight_vars}{A character vector with the names of count variables that make up the weights}

\item{values}{A numeric vector of known constants associated with each count}
}
\value{
A new weighted average variable that can be added as a new column to the DP data frame
}
\description{
This function that adds a weighted aaveraage column to dp data with appropriate standard deviation in first row for use in lmdp
}
